refactor(e2e-tests): Distribute `types.ts` to corresponding modules
