+++ apps/e2e-tests/src/utils/services/types.ts
@@ -1,0 +1,55 @@
+import { Schema as S } from 'effect';
+
+import type { FeedResult } from '../services/generate-signature';
+
+export const UpdatesToNetworkMetric = S.Struct({
+  name: S.Literal('updates_to_networks'),
+  type: S.Literal('COUNTER'),
+  metrics: S.Array(
+    S.Struct({
+      value: S.Number,
+      labels: S.Struct({
+        Network: S.String,
+        FeedId: S.String,
+      }),
+    }),
+  ),
+});
+
+export type UpdatesToNetwork = Record<string, Record<string, number>>;
+
+const Numerical = S.Struct({
+  Numerical: S.Number,
+});
+
+export const FeedAggregateHistorySchema = S.Struct({
+  aggregate_history: S.Record({
+    key: S.String,
+    value: S.Array(
+      S.Struct({
+        value: Numerical,
+        update_number: S.Number,
+        end_slot_timestamp: S.Number,
+      }),
+    ),
+  }),
+});
+
+export type FeedAggregateHistory = typeof FeedAggregateHistorySchema.Type;
+
+export type ReportData = {
+  feed_id: string;
+  value: FeedResult;
+};
+
+type ReportPayloadData = {
+  feed_id: string;
+  reporter_id: number;
+  timestamp: number;
+  signature: string;
+};
+
+export type ReportPayload = {
+  payload_metadata: ReportPayloadData;
+  result: FeedResult;
+};
